Jira plugin: add support for Components and Google Chat Threads (#51)
This PR contains two changes in one. Please let me know if you'd prefer to split them up.

* Adds support for monitoring components
Some Jira projects use components to track smaller parts. This allows the bot to monitor for those components, and send the notifications to chat rooms/channels which are specific for them.

* Add support for Google Chat threads
Originally each time the bot sends a self-initiated message to a Google Chat room it initializes a new thread. This patch adds support for sending notifications to the same thread. The only caveat is that, because of Google Chat API limitations, the thread needs to pre-exist (that's reflected in the documentation).

I tried to make this semi-independent of the protocol being used, such that it can be modified to use threads in other outputs like Slack, however, right now I don't have access to other chat tools that use threads to test.

NOTE: in the patch, you will see some unrelated lines of code being changed. That's because I use Visual Studio Code, and it has the habit of trying to re-format surrounding code when you start making changes. Sorry for that.